support emit code model
