+++ packages/typespec-client-generator-core/src/interfaces.ts
@@ -75,6 +75,7 @@ export interface SdkEmitterOptions {
    */
   "examples-directory"?: string;
   "examples-dir"?: string;
+  "emitter-name"?: string;
 }
 
 export interface SdkClient {
