+++ packages/typespec-client-generator-core/test/decorators.test.ts
@@ -3,7 +3,6 @@ import { expectDiagnostics } from "@typespec/compiler/testing";
 import { deepStrictEqual, ok, strictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
 import {
-  createSdkContext,
   getAccess,
   getClient,
   getClientNameOverride,
@@ -2191,40 +2190,6 @@ describe("typespec-client-generator-core: decorators", () => {
       strictEqual(aOps.length, 1);
       a = aOps.find((x) => x.name === "a");
       ok(a);
-    });
-  });
-
-  describe("createSdkContext", () => {
-    it("multiple call with versioning", async () => {
-      const tsp = `
-        @service({
-          title: "Contoso Widget Manager",
-        })
-        @versioned(Contoso.WidgetManager.Versions)
-        namespace Contoso.WidgetManager;
-        
-        enum Versions {
-          v1,
-        }
-
-        @client({name: "TestClient"})
-        @test
-        interface Test {}
-      `;
-
-      const runnerWithVersion = await createSdkTestRunner({
-        emitterName: "@azure-tools/typespec-python",
-      });
-
-      await runnerWithVersion.compile(tsp);
-      let clients = listClients(runnerWithVersion.context);
-      strictEqual(clients.length, 1);
-      ok(clients[0].type);
-
-      const newSdkContext = await createSdkContext(runnerWithVersion.context.emitContext);
-      clients = listClients(newSdkContext);
-      strictEqual(clients.length, 1);
-      ok(clients[0].type);
     });
   });
