+++ packages/typespec-client-generator-core/test/public-utils.test.ts
@@ -24,12 +24,7 @@ import {
   isAzureCoreModel,
 } from "../src/public-utils.js";
 import { getAllModels, getSdkUnion } from "../src/types.js";
-import {
-  SdkTestRunner,
-  createSdkContextTestHelper,
-  createSdkTestRunner,
-  createTcgcTestRunnerForEmitter,
-} from "./test-host.js";
+import { SdkTestRunner, createSdkContextTestHelper, createSdkTestRunner } from "./test-host.js";
 
 describe("typespec-client-generator-core: public-utils", () => {
   let runner: SdkTestRunner;
@@ -402,7 +397,7 @@ describe("typespec-client-generator-core: public-utils", () => {
   describe("getPropertyNames", () => {
     it("property language projected name", async () => {
       async function helper(emitterName: string, expectedLibraryName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -428,7 +423,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("property language projected name augmented", async () => {
       async function helper(emitterName: string, expectedLibraryName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -453,7 +448,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("property client projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -475,7 +470,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("property no projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -494,7 +489,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("property with projected client and json name", async () => {
       async function helper(emitterName: string, expectedLibraryName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -521,7 +516,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("property with projected language and json name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -544,7 +539,7 @@ describe("typespec-client-generator-core: public-utils", () => {
   describe("getLibraryName", () => {
     it("operation client projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { func } = (await runner.compile(`
         @test @clientName("rightName") op func(@query("api-version") myApiVersion: string): void;
       `)) as { func: Operation };
@@ -557,7 +552,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("operation language projected name", async () => {
       async function helper(emitterName: string, expected: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { func } = (await runner.compile(`
         @test
         @clientName("madeForCS", "csharp")
@@ -575,7 +570,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("operation language projected name augmented", async () => {
       async function helper(emitterName: string, expected: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { func } = (await runner.compile(`
         @test
         op func(@query("api-version") myApiVersion: string): void;
@@ -594,7 +589,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("operation json projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { func } = (await runner.compile(`
         @test
         @encodedName("application/json", "NotToUseMeAsName") // Should be ignored
@@ -609,7 +604,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("operation no projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { func } = (await runner.compile(`
         @test
         op func(@query("api-version") myApiVersion: string): void;
@@ -623,7 +618,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model client projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         @clientName("RightName")
@@ -640,7 +635,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model language projected name", async () => {
       async function helper(emitterName: string, expected: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         @clientName("CsharpModel", "csharp")
@@ -660,7 +655,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model language projected name augmented", async () => {
       async function helper(emitterName: string, expected: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -681,7 +676,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model json projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         @encodedName("application/json", "NotToUseMeAsName") // Should be ignored
@@ -698,7 +693,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model no projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -714,7 +709,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model friendly name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         @friendlyName("FriendlyName")
@@ -731,7 +726,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("model friendly name augmented", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         model MyModel {
@@ -749,7 +744,7 @@ describe("typespec-client-generator-core: public-utils", () => {
 
     it("should return language specific name when both language specific name and friendly name exist", async () => {
       async function helper(expected: string, emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         @friendlyName("FriendlyName")
@@ -771,7 +766,7 @@ describe("typespec-client-generator-core: public-utils", () => {
 
     it("should return client name when both client name and friendly name exist", async () => {
       async function helper(expected: string, emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { MyModel } = (await runner.compile(`
         @test
         @friendlyName("FriendlyName")
@@ -790,7 +785,7 @@ describe("typespec-client-generator-core: public-utils", () => {
 
     it("parameter client projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { param } = (await runner.compile(`
         op func(
           @test
@@ -808,7 +803,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("parameter language projected name", async () => {
       async function helper(emitterName: string, expected: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { param } = (await runner.compile(`
         op func(
           @test
@@ -830,7 +825,7 @@ describe("typespec-client-generator-core: public-utils", () => {
 
     it("parameter json projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { param } = (await runner.compile(`
         op func(
           @test
@@ -848,7 +843,7 @@ describe("typespec-client-generator-core: public-utils", () => {
     });
     it("parameter no projected name", async () => {
       async function helper(emitterName: string) {
-        const runner = await createTcgcTestRunnerForEmitter(emitterName);
+        const runner = await createSdkTestRunner({ emitterName });
         const { param } = (await runner.compile(`
         op func(
           @test
